+++ src/components/entry/EntryContainer.tsx
@@ -76,7 +76,7 @@ const EntryContainer = ({
     <>
       <div
         ref={scrollRef}
-        className='flex flex-col flex-1 overflow-auto'
+        className='flex flex-col flex-1 overflow-auto pt-14 lg:pt-0'
         onScroll={() =>
           handleScroll(
             scrollRef,