+++ src/components/feed/FeedMenu.tsx
@@ -26,20 +26,22 @@ const FeedMenu = ({
       <CircleLoader color={STUColor} size={50} />
     </div>
   ) : feeds.length === 0 ? (
-    <div>{t('modal.feedMenu.empty')}</div>
+    <div className='flex h-40 items-center justify-center text-center text-xl font-extrabold'>
+      {t('modal.feedMenu.empty')}
+    </div>
   ) : (
     <>
-      <div className='flex w-full flex-wrap'>
+      <div className='flex w-full flex-wrap mb-2'>
         {activeFeeds.map((feeds, index) => (
           <div
             key={index}
-            className={`${
-              searchBar ? 'w-1/2' : 'w-full sm:w-1/2 md:w-1/3 lg:w-1/4'
-            } p-2`}
+            className={
+              searchBar ? 'w-full' : 'w-full sm:w-1/2 md:w-1/3 lg:w-1/4'
+            }
           >
             <button
               type='button'
-              className='bg-STUColor text-sm hover:bg-red w-full h-full p-2 flex justify-between items-center text-white rounded-md'
+              className='bg-STUColor p-2 text-sm hover:bg-red w-full h-full flex justify-between items-center text-white rounded-md'
               onClick={() =>
                 setActiveFeeds(
                   // Return only those which id does not equal
@@ -61,7 +63,7 @@ const FeedMenu = ({
             key={index}
             className={`${
               searchBar ? 'w-1/2' : 'w-full sm:w-1/2 md:w-1/3 lg:w-1/4'
-            } p-2`}
+            } p-1`}
           >
             <button
               type='button'