+++ src/pages/admin/entries/AdminEditEntry.tsx
@@ -2,7 +2,7 @@ import Breadcrumb from '../../../components/common/Breadcrumb';
 
 const AdminEditEntry = () => {
   return (
-    <div className='flex-1'>
+    <div className='flex-1 p-4 pt-14 lg:pt-4'>
       <Breadcrumb />
       <div>edit</div>
     </div>